restructure back to original so that autocomplete works and layout stays the same
